requirements.txt
altgraph==0.17.4
et_xmlfile==2.0.0
joblib==1.5.2
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
Pygments==2.19.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32-ctypes==0.2.3
rich==14.2.0
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
